refactor: re-organize code
