+++ crd-generator/src/main/java/io/fabric8/crd/generator/utils/Strings.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.utils;
+package io.fabric8.crd.generator.utils;
 
 public class Strings {
