+++ crd-generator/src/main/java/io/fabric8/crd/generator/utils/Types.java
