+++ crd-generator/src/main/java/io/fabric8/crd/generator/v1/decorator/AddCustomResourceDefinitionResourceDecorator.java
