+++ crd-generator/src/main/java/io/fabric8/crd/generator/v1beta1/JsonSchema.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.crd.v1beta1;
+package io.fabric8.crd.generator.v1beta1;
 
-import io.fabric8.crd.util.Types;
+import io.fabric8.crd.generator.utils.Types;
 import io.fabric8.kubernetes.api.model.apiextensions.v1beta1.JSONSchemaProps;
 import io.fabric8.kubernetes.api.model.apiextensions.v1beta1.JSONSchemaPropsBuilder;
 import io.fabric8.kubernetes.client.CustomResource;
