+++ crd-generator/src/main/java/io/fabric8/crd/generator/v1beta1/decorator/AddStatusSubresourceDecorator.java
