+++ crd-generator/src/main/java/io/fabric8/crd/generator/visitor/SpecReplicasPathDetector.java
