+++ crd-generator/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-io.fabric8.crd.apt.CustomResourceAnnotationProcessor
+io.fabric8.crd.generator.apt.CustomResourceAnnotationProcessor
