+++ crd-generator/src/test/java/io/fabric8/crd/generator/utils/TypesTest.java
